requirements/dev.txt
asgiref==3.7.2
asttokens==2.4.1
decorator==5.1.1
Django==5.0.3
django-debug-toolbar==4.3.0
django-modeltranslation==0.18.12
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
exceptiongroup==1.2.0
executing==2.0.1
gunicorn==21.2.0
ipython==8.22.2
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==24.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.10.0
wcwidth==0.2.13
